drivers: mfd: senoko: Add Senoko device
Senoko is the power supply board present on some Kosagi boards.  It
acts as a combination RTC, power supply, keyboard, and GPIO board.  This
driver implements some of these features.

Signed-off-by: Sean Cross <xobs@kosagi.com>

(Jookia: Ported to v4.14 and merged with commits below:)

senoko: repeatedly try to powerdown

If the bus is busy, the powerdown message will fail.  Bus contention can
cause the STM32 I2C block to lock up.

To work around this, we have a timer on the Senoko that resets I2C if it's
busy for more than a certain amount of time.

If we fail to power off. sleep an ever-increasing amount of time in order
to give the Senoko board a chance to revive itself in this case.

Signed-off-by: Sean Cross <xobs@kosagi.com>

drivers: power: Add senoko power supply

Senoko power supply can detect the state of the AC connection.

Signed-off-by: Sean Cross <xobs@kosagi.com>

(Jookia: Moved to supply subdirectory)

drivers: input: keyboard: Add Senoko power button

Senoko acts like a keyboard with a single power button.  This button
supports wake-from-suspend.

Signed-off-by: Sean Cross <xobs@kosagi.com>