novena: drm: imx: HACK: Manually specify CRTCs
The CRTC assignment code doesn't realize that ldb needs two CRTCs that
must be a part of the same IPU when using both channels.

This patch hacks it so that the "IPU" only uses IPU1, and the DW-HDMI code
only uses IPU2.  It is Novena-specific and should not be used anywhere else.

Signed-off-by: Sean Cross <xobs@kosagi.com>