fs: pstore: eepromoops: Fix compatibility with 4.19
The pstore module has undergone some changes.

It now uses 64-bit timestamps, so convert between the 32-bit used by the
existing eepromoops.

The spinlock has been replaced with a semaphore, so remove the call like
the RAM module has done.

The pstore initialization code requires the pstore_info.flags value to
be set, so set it to the value that gets oops and panic dumps.