ARM: dts: imx6q-novena: Always enable the es8328-power regulator
Linux helpfully shuts down the es8328-power regulator if the es8328
audio codec isn't in use. Unhelpfully this seems to kill the I2C3 bus
and break a variety of functions such as the screen, EEPROM and Senoko.

Keep the regular enabled even if it's not used by the es8328 codec.